update :fix bug#64812

+++ plugins/security-updates/upgrade/src/updatedbus.cpp
@@ -261,7 +261,7 @@ void UpdateDbus::getAppMessageSignal(QMap<QString, QVariant> map, QStringList ur
             msg.name = nameList.at(i);
             msg.fullname = fullnameList.at(i);
             QString size = sizeList.at(i);
-            msg.size = size.toInt();
+            msg.size = size.toLong();
             appAllMsg.msg.depList.append(msg);
         }
     }
